+++ database/seeders/IntakeTemplateSeeder.php
@@ -25,6 +25,7 @@ class IntakeTemplateSeeder extends Seeder
             database_path('data/templates/airco/v8.php'),
             database_path('data/templates/airco/v9.php'),
             database_path('data/templates/airco/v10.php'),
+            database_path('data/templates/airco/v11.php'),
         ];
 
         foreach ($configs as $path) {